Test de aeropuertos iguales reparado. El problema estaba en el boolean de validarAeropuerto, porque integer es objeto y estaba comparado mediando simbolos de primitivos. Ser agrega intValue y SOLUCIONADOOOO

+++ src/main/java/ar/com/ada/api/aladas/services/VueloService.java
@@ -80,7 +80,7 @@ public class VueloService {
          * if(vuelo.getAeropuertoDestino() != vuelo.getAeropuertoOrigen()) return true;
          * else return false;
          */
-        return vuelo.getAeropuertoDestino() != vuelo.getAeropuertoOrigen();
+        return vuelo.getAeropuertoDestino().intValue() != vuelo.getAeropuertoOrigen().intValue();
 
     }
